Moved item_callbacks.py and inventory.py to helper_functions package
